# StablingApi/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "StablingApi",
    "description": "API ИС поддержки управления КСК",
    "version": "v1"
  },
  "paths": {
...
    "/api/Trainings": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_GetAll",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Training"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "training",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Training"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Training"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_Update",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "training",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Training"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Training"
            }
          }
        }
      }
    },
    "/api/Trainings/{id}": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Training"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/Trainings/ForShow/{id}": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_GetForShow",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/TrainingForShow"
            }
          }
        }
      }
    },
    "/api/Trainings/ByDay/{dateTime}": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_GetByDay",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "dateTime",
            "in": "path",
            "required": true,
            "format": "date-time",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Training"
              }
            }
          }
        }
      }
    },
    "/api/Trainings/ByWeek/{dateTime}": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_GetByWeek",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "dateTime",
            "in": "path",
            "required": true,
            "format": "date-time",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Training"
              }
            }
          }
        }
      }
    },
...
    "/api/Trainings/ByHorseDay/{dateTime}/{horseId}": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_GetByHorseDay",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "dateTime",
            "in": "path",
            "required": true,
            "format": "date-time",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "horseId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Training"
              }
            }
          }
        }
      }
    },
    "/api/Trainings/ForShow/ByDate/{dateTime}": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_GetForShowByDate",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "dateTime",
            "in": "path",
            "required": true,
            "format": "date-time",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TrainingForShow"
              }
            }
          }
        }
      }
    },
    "/api/Trainings/PaidStatus/{trainingId}": {
      "put": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_ChangePaidStatus",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "trainingId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
    "/api/Trainings/Transfer/{id}": {
      "put": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_Transfer",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "dateTime",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Training"
            }
          }
        }
      }
    },
    "/api/Trainings/Types": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_GetTypes",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TrainingType"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_Create2",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TrainingType"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/TrainingType"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_Update2",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TrainingType"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/TrainingType"
            }
          }
        }
      }
    },
    "/api/Trainings/Types/{typeId}": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "operationId": "Trainings_GetType",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "typeId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/TrainingType"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Training": {
      "type": "object",
      "required": [
        "trainingId",
        "trainingTypeId",
        "trainerId",
        "horseId",
        "clientId",
        "trainingDateTime",
        "duration",
        "isPaid"
      ],
      "properties": {
        "trainingId": {
          "type": "integer",
          "format": "int32"
        },
        "trainingTypeId": {
          "type": "integer",
          "format": "int32"
        },
        "trainerId": {
          "type": "integer",
          "format": "int32"
        },
        "horseId": {
          "type": "integer",
          "format": "int32"
        },
        "clientId": {
          "type": "integer",
          "format": "int32"
        },
        "trainingDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "duration": {
          "type": "number",
          "format": "decimal"
        },
        "isPaid": {
          "type": "boolean"
        }
      }
    },
    "TrainingForShow": {
      "type": "object",
      "required": [
        "trainingId",
        "trainingTypeId",
        "typeName",
        "trainerId",
        "trainerName",
        "horseId",
        "horseName",
        "clientId",
        "clientName",
        "trainingStart",
        "trainingFinish",
        "isPaid",
        "trainingLabel",
        "trainingFunds"
      ],
      "properties": {
        "trainingId": {
          "type": "integer",
          "format": "int32"
        },
        "trainingTypeId": {
          "type": "integer",
          "format": "int32"
        },
        "typeName": {
          "type": "string"
        },
        "trainerId": {
          "type": "integer",
          "format": "int32"
        },
        "trainerName": {
          "type": "string"
        },
        "horseId": {
          "type": "integer",
          "format": "int32"
        },
        "horseName": {
          "type": "string"
        },
        "clientId": {
          "type": "integer",
          "format": "int32"
        },
        "clientName": {
          "type": "string"
        },
        "trainingStart": {
          "type": "string",
          "format": "date-time"
        },
        "trainingFinish": {
          "type": "string",
          "format": "date-time"
        },
        "isPaid": {
          "type": "boolean"
        },
        "trainingLabel": {
          "type": "string"
        },
        "trainingFunds": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "TrainingType": {
      "type": "object",
      "required": [
        "trainingTypeId",
        "typeName",
        "typePrice",
        "typeDuration",
        "isAvailable"
      ],
      "properties": {
        "trainingTypeId": {
          "type": "integer",
          "format": "int32"
        },
        "typeName": {
          "type": "string"
        },
        "typePrice": {
          "type": "integer",
          "format": "int32"
        },
        "typeDuration": {
          "type": "number",
          "format": "decimal"
        },
        "isAvailable": {
          "type": "boolean"
        }
      }
    }
  }
}